+++ MMM-Fuel.js
@@ -52,6 +52,7 @@ Module.register('MMM-Fuel', {
      * @property {boolean} showAddress - Flag to show the gas stations address.
      * @property {boolean} showOpenOnly - Flag to show only open gas stations or all.
      * @property {boolean} showDistance - Flag to show the distance to your specified position.
+     * @property {boolean} showBrand - Flag to show the brand instead of the name.
      * @property {boolean} iconHeader - Flag to display the car icon in the header.
      * @property {boolean} rotate - Flag to enable/disable rotation between sort by price and distance.
      * @property {string[]} types - Fuel types to show.
@@ -74,6 +75,7 @@ Module.register('MMM-Fuel', {
         showAddress: true,
         showOpenOnly: false,
         showDistance: true,
+        showBrand: false,
         iconHeader: true,
         rotate: true,
         types: ['diesel'],